YouTube-Summarizer/pyproject.toml
[tool.poetry]
name = "youtube-summarizer"
version = "0.1.0"
description = ""
authors = ["jkusnierz01 <jedrek33@onet.pl>"]
readme = "README.md"
package-mode = false  # Disable packaging mode

[tool.poetry.dependencies]
python = "^3.11"
numpy = "<2"
nltk = "3.6.5"
accelerate = "^0.33.0"
websockets = "^13.0"
torch = "^2.4.1"
huggingface-hub = "^0.24.7"
pydub = "^0.25.1"


[tool.poetry.group.fastapi.dependencies]
openai-whisper = {git = "https://github.com/openai/whisper.git"}
uuid = "^1.30"
yt-dlp = "^2024.8.6"
fastapi = "0.110.1"
uvicorn = "^0.22.0"
pyannote-audio = "^3.3.2"

[tool.poetry.group.reflex.dependencies]
reflex = "^0.5.9"
reflex-motion = "^0.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
